+++ docs/CHANGE_LOG.md
@@ -3,7 +3,7 @@
 ## Version 0.5.0:
 
 #### **New Features:**
-- Browser URL parameters support! This will allow you to force Journey.js to open when the page has finished loading, and state what step to start on.
+- Browser URL parameters support! This will allow you to force Journey.js to open when the page has finished loading and state what step to start on.
 
 #### **Configuration Options:**
 - Added a new configuration option "browserUrlParametersEnabled" (defaults to true), which states if browser URL parameter reading support is enabled.