+++ docs/CHANGE_LOG.md
@@ -6,6 +6,9 @@
 - Added a new binding option called "alignTop", which states if the dialog should be aligned to the top of the element.
 - Added a new binding option called "alignRight", which states if the dialog should be aligned to the right of the element.
 
+#### **Binding Options - Custom Triggers:**
+- Added a new binding custom trigger called "onOpen", which states an event that should be triggered when the dialog is first opened.
+
 #### **Documentation:**
 - Added examples to all documentation areas.
 - Reorganized the documentation for the project.